+++ service/d-installer.gemspec
