+++ service/lib/dinstaller.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
 # Specific YaST code lives within this module.
-module Yast2
+module DInstaller
 end
-
-require "yast2/installer"
