+++ service/lib/dinstaller/dbus.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+# Namespace for DBus API
+module DInstaller
+  module DBus
+  end
+end
+
+require "dinstaller/dbus/manager"
+require "dinstaller/dbus/language"
+require "dinstaller/dbus/software"
